clean up objconv API
