+++ emit.go
@@ -61,96 +61,6 @@ type Emitter interface {
 	EmitMapNext() error
 }
 
-// ValueEmitter is a special kind of emitter, instead of serializing the values
-// it receives it builds an in-memory representation of the data.
-//
-// This is useful for testing the high-level API of the package without actually
-// having to generate a serialized representation.
-type ValueEmitter struct {
-	stack []interface{}
-	marks []int
-}
-
-// NewValueEmitter returns a pointer to a new ValueEmitter object.
-func NewValueEmitter() *ValueEmitter {
-	return &ValueEmitter{}
-}
-
-// Value returns the value built in the emitter.
-func (e *ValueEmitter) Value() interface{} { return e.stack[0] }
-
-func (e *ValueEmitter) EmitNil() error { return e.push(nil) }
-
-func (e *ValueEmitter) EmitBool(v bool) error { return e.push(v) }
-
-func (e *ValueEmitter) EmitInt(v int64, _ int) error { return e.push(v) }
-
-func (e *ValueEmitter) EmitUint(v uint64, _ int) error { return e.push(v) }
-
-func (e *ValueEmitter) EmitFloat(v float64, _ int) error { return e.push(v) }
-
-func (e *ValueEmitter) EmitString(v string) error { return e.push(v) }
-
-func (e *ValueEmitter) EmitBytes(v []byte) error { return e.push(v) }
-
-func (e *ValueEmitter) EmitTime(v time.Time) error { return e.push(v) }
-
-func (e *ValueEmitter) EmitDuration(v time.Duration) error { return e.push(v) }
-
-func (e *ValueEmitter) EmitError(v error) error { return e.push(v) }
-
-func (e *ValueEmitter) EmitArrayBegin(v int) error { return e.pushMark() }
-
-func (e *ValueEmitter) EmitArrayEnd() error {
-	v := e.pop(e.popMark())
-	a := make([]interface{}, len(v))
-	copy(a, v)
-	return e.push(a)
-}
-
-func (e *ValueEmitter) EmitArrayNext() error { return nil }
-
-func (e *ValueEmitter) EmitMapBegin(v int) error { return e.pushMark() }
-
-func (e *ValueEmitter) EmitMapEnd() error {
-	v := e.pop(e.popMark())
-	n := len(v)
-	m := make(map[interface{}]interface{}, n/2)
-
-	for i := 0; i != n; i += 2 {
-		m[v[i]] = v[i+1]
-	}
-
-	return e.push(m)
-}
-
-func (e *ValueEmitter) EmitMapValue() error { return nil }
-
-func (e *ValueEmitter) EmitMapNext() error { return nil }
-
-func (e *ValueEmitter) push(v interface{}) error {
-	e.stack = append(e.stack, v)
-	return nil
-}
-
-func (e *ValueEmitter) pop(n int) []interface{} {
-	v := e.stack[n:]
-	e.stack = e.stack[:n]
-	return v
-}
-
-func (e *ValueEmitter) pushMark() error {
-	e.marks = append(e.marks, len(e.stack))
-	return nil
-}
-
-func (e *ValueEmitter) popMark() int {
-	n := len(e.marks) - 1
-	m := e.marks[n]
-	e.marks = e.marks[:n]
-	return m
-}
-
 type discardEmitter struct{}
 
 func (e discardEmitter) EmitNil() error                     { return nil }
