+++ objutil/empty.go
@@ -1,0 +1,45 @@
+package objutil
+
+// Copyright 2009 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+import (
+	"reflect"
+	"unsafe"
+)
+
+// IsEmpty returns true if the value given as argument would be considered
+// empty by the standard library packages, and therefore not serialized if
+// `omitempty` is set on a struct field with this value.
+func IsEmpty(v interface{}) bool {
+	return IsEmptyValue(reflect.ValueOf(v))
+}
+
+// IsEmptyValue returns true if the value given as argument would be considered
+// empty by the standard library packages, and therefore not serialized if
+// `omitempty` is set on a struct field with this value.
+//
+// Based on https://golang.org/src/encoding/json/encode.go?h=isEmpty
+func IsEmptyValue(v reflect.Value) bool {
+	if !v.IsValid() {
+		return true // nil interface{}
+	}
+	switch v.Kind() {
+	case reflect.Array, reflect.Map, reflect.Slice, reflect.String:
+		return v.Len() == 0
+	case reflect.Bool:
+		return !v.Bool()
+	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
+		return v.Int() == 0
+	case reflect.Uint, reflect.Uint8, reflect.Uint16, reflect.Uint32, reflect.Uint64, reflect.Uintptr:
+		return v.Uint() == 0
+	case reflect.Float32, reflect.Float64:
+		return v.Float() == 0
+	case reflect.Interface, reflect.Ptr, reflect.Chan, reflect.Func:
+		return v.IsNil()
+	case reflect.UnsafePointer:
+		return unsafe.Pointer(v.Pointer()) == nil
+	}
+	return false
+}
